+++ packages/connect/src/webrtc/connection.ts
@@ -123,7 +123,8 @@ class WebRTCConnection implements MultiaddrConnection<StreamType> {
             yield* getAbortableSource(toU8aStream(source), this.options?.signal)
             deferred.resolve()
           } catch (err: any) {
-            if (err.type === 'aborted') {
+            if (err.type === 'aborted' || err.code === 'ABORT_ERR') {
+              // We can safely ignore abort errors
               deferred.resolve()
             } else {
               this.error(`sink error thrown`, err.message)