+++ packages/core-ethereum/src/ethereum.ts
@@ -61,7 +61,7 @@ export async function createChainWrapper(
     wallet
   ) as HoprChannels
 
-  const genesisBlock = (await provider.getTransaction(hoprChannelsDeployment.transactionHash)).blockNumber
+  const genesisBlock = parseInt(hoprChannelsDeployment.blockNumber)
   const channelClosureSecs = await channels.secsClosure()
 
   const transactions = new TransactionManager()
@@ -489,6 +489,7 @@ export async function createChainWrapper(
       token.removeAllListeners()
     },
     getChannels: () => channels,
+    getToken: () => token,
     getPrivateKey: () => utils.arrayify(wallet.privateKey),
     getPublicKey: () => PublicKey.fromString(utils.computePublicKey(wallet.publicKey, true)),
     getInfo: () => ({