+++ packages/core-ethereum/src/index.ts
@@ -113,7 +113,7 @@ export default class HoprCoreEthereum extends EventEmitter {
 
         const hoprBalance = await this.chain.getBalance(this.publicKey.toAddress())
         await this.db.setHoprBalance(hoprBalance)
-        log(`set HOPR balance to ${hoprBalance.toFormattedString()}`)
+        log(`set own HOPR balance to ${hoprBalance.toFormattedString()}`)
 
         await this.indexer.start(this.chain, this.chain.getGenesisBlock())